% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph2json.R
\name{igraph2json}
\alias{igraph2json}
\title{Convert igraph network to networkx Graph}
\usage{
igraph2json(r_igraph, path)
}
\arguments{
\item{r_igraph}{An \code{igraph} \code{graph} object}

\item{path}{An file \code{path} to save to \code{Networkx} \code{json} format}
}
\description{
Convert \code{igraph} network to a \code{networkx} python object
including Vertex and Edge attributes and adding \code{igraph_id}
and \code{igraph_name} to retain \code{igraph} indexing
}
\examples{
\dontrun{
example_bi_modal_graph <- igraph::sample_bipartite(4, 5, p = .5)
igraph2json(example_bi_modal_graph, "test.json")
loaded_bi_modal_graph <- netx2igraph("test.json")
as_edgelist(example_bi_modal_graph) == as_edgelist(loaded_bi_modal_graph)
}
}
