% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python-dependencies.R
\name{install_networkx}
\alias{install_networkx}
\title{Install the python networkx library}
\usage{
install_networkx(method = "auto", conda = "auto")
}
\arguments{
\item{method}{string passed to reticulate::py_install 'method' parameter}

\item{conda}{string passed to reticulate::py_install 'conda' parameter}
}
\description{
This is an option for installing networkx if it isn't added automatically
}
