% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python-dependencies.R, R/zzz.R
\name{install_pandas}
\alias{install_pandas}
\title{Install the python pandas library}
\usage{
install_pandas(method = "auto", conda = "auto")

install_pandas(method = "auto", conda = "auto")
}
\arguments{
\item{method}{string passed to reticulate::py_install 'method' parameter}

\item{conda}{string passed to reticulate::py_install 'conda' parameter}
}
\description{
This is an option for installing pandas if it isn't added automatically

This is an option for installing pandas if it isn't added automatically
}
