% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python-dependencies.R, R/zzz.R
\docType{data}
\name{networkx}
\alias{networkx}
\title{Placeholder for the python networkx package}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.

An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
networkx

networkx
}
\description{
This provides a namespace placeholder for loading the python networkx library

This provides a namespace placeholder for loading the python networkx library
}
\keyword{datasets}
