% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netx2igraph.R
\name{netx2igraph}
\alias{netx2igraph}
\title{Convert networkx graph to an igraph Graph}
\usage{
netx2igraph(path)
}
\arguments{
\item{path}{A path to a \code{networkx} json file to load.}
}
\value{
An \code{igraph} object equivalent with all the components in
\code{path}.
}
\description{
This uses \code{networkx} and \code{pandas} python libraries to convert
a \code{networkx} graph object to an R \code{igraph} object, including
whether the graph is bipartite.
}
\examples{
\dontrun{
library(igraph)
board_network <- netx2igraph("path/to/networkx/board-network.json")
vcount(board_network)
}
}
