% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python-dependencies.R
\docType{data}
\name{pandas}
\alias{pandas}
\title{Placeholder for the python pandas package}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
pandas
}
\description{
This provides a namespace placeholder for loading the python pandas library
}
\keyword{datasets}
