% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-python-dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install all needed python dependencies}
\usage{
install_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{method}{string passed to reticulate::py_install 'method' parameter}

\item{conda}{string passed to reticulate::py_install 'conda' parameter}
}
\description{
Install python dependencies not currently in the loaded in the current
reticulate python environment
}
